LayoutTests/fast/css/getPropertyValue-columns-expected.txt

Bug 111011: getPropertyValue for columns returns null, should compute the shorthand value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS columnsValue("columns1") is "10px"
PASS columnsValue("columns2") is "10"
PASS columnsValue("columns3") is "10px"
PASS columnsValue("columns4") is "auto"
PASS columnsValue("columns5") is "2"
PASS columnsValue("columns6") is "10px 2"
PASS columnsValue("columns7") is "auto auto"
NOTE: 'foo' is an illegal CSS value for 'columns'.
PASS columnsValue("columns8") is ""
NOTE: If only few longhand properties are specified, getPropertyValue for shorthand property returns the empty string.
PASS columnsValue("columns9") is ""
PASS successfullyParsed is true

TEST COMPLETE
